requirements.txt
aiogram>=3.22,<4
python-dotenv>=1.0
gspread>=5.12
google-auth>=2.34
